+++ lime-loader/src/test/java/com/here/gluecodium/validator/LimeInterfacesValidatorTest.kt
@@ -1,0 +1,73 @@
+/*
+ * Copyright (C) 2016-2020 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package com.here.gluecodium.validator
+
+import com.here.gluecodium.model.lime.LimeAttributeType
+import com.here.gluecodium.model.lime.LimeAttributes
+import com.here.gluecodium.model.lime.LimeElement
+import com.here.gluecodium.model.lime.LimeInterface
+import com.here.gluecodium.model.lime.LimeModel
+import com.here.gluecodium.model.lime.LimePath.Companion.EMPTY_PATH
+import io.mockk.mockk
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertTrue
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.junit.runners.JUnit4
+
+@RunWith(JUnit4::class)
+class LimeInterfacesValidatorTest {
+
+    private val allElements = mutableMapOf<String, LimeElement>()
+    private val limeModel = LimeModel(allElements, emptyList())
+
+    private val validator = LimeInterfacesValidator(mockk(relaxed = true))
+
+    @Test
+    fun validateInterfaceWithoutAttributes() {
+        allElements[""] = LimeInterface(EMPTY_PATH)
+
+        assertTrue(validator.validate(limeModel))
+    }
+
+    @Test
+    fun validateInterfaceWithPointerEquatable() {
+        allElements[""] = LimeInterface(
+            EMPTY_PATH,
+            attributes = LimeAttributes.Builder()
+                .addAttribute(LimeAttributeType.POINTER_EQUATABLE)
+                .build()
+        )
+
+        assertFalse(validator.validate(limeModel))
+    }
+
+    @Test
+    fun validateInterfaceWithEquatable() {
+        allElements[""] = LimeInterface(
+            EMPTY_PATH,
+            attributes = LimeAttributes.Builder()
+                .addAttribute(LimeAttributeType.EQUATABLE)
+                .build()
+        )
+
+        assertTrue(validator.validate(limeModel))
+    }
+}