+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Breaking changes:
+  * Added validation against using `@PointerEquatable` attribute on interfaces in IDL files. This
+    usage is currently unsupported but was erroneously allowed before.
+
 ## 6.3.2
 Release date: 2020-03-10
 ### Bug fixes:
@@ -85,7 +90,6 @@ Release date: 2019-11-21
 ### Bug fixes:
   * Fixed C++ compilation issue for `@Equatable` structs with `Date` type fields.
   * Fixed Java compilation issue for interfaces with same name in different packages.
-
 ### Breaking changes:
   * Removed "stdout" command line option.